feat: added a fourth table for select purchase ticket
