+++ frontend/dwaves-app/src/components/index.ts
@@ -16,7 +16,8 @@ import { CreatePlaylist } from "components/CreatePlaylist";
 import { UserPlaylists } from "./UserPlaylists";
 import { PlaylistContent } from "./PlaylistContent";
 import { CreateConcert } from "./CreateConcert";
-import { NftTickets } from './NftTickets'
+import { YourNftTickets } from "./YourNftTickets";
+import { TicketPurchase } from "./TicketPurchase";
 import {YourConcerts} from "./YourConcerts";
 import { Kitsune } from "./Kitsune";
 import { Footer } from "components/Footer";
@@ -40,7 +41,8 @@ export {
   UserPlaylists,
   PlaylistContent,
   CreateConcert,
-  NftTickets,
+  YourNftTickets,
+  TicketPurchase,
   YourConcerts,
   Kitsune,
   Footer
