+++ frontend/dwaves-app/src/views/Marketplace.tsx
@@ -2,7 +2,8 @@ import React, { useEffect, useState } from 'react'
 import {
   CreateConcert,
   Kitsune,
-  NftTickets,
+  TicketPurchase,
+  YourNftTickets,
   SwitchTab,
   YourConcerts,
 } from '../components'
@@ -28,7 +29,7 @@ export const Marketplace: React.FC<Props> = ({
   requestConnectionMetamask,
 }) => {
   const [myUsername, setMyUsername] = useState('')
-  const [showTickets, setShowTickets] = useState('NFT Tickets')
+  const [showTickets, setShowTickets] = useState('Ticket Counter')
   const [balance, setBalance] = useState('')
   const [chainName, setChainName] = useState('')
   const [chainId, setChainId] = useState('')
@@ -193,8 +194,11 @@ export const Marketplace: React.FC<Props> = ({
         artist: ticket[6],
         price: parseFloat(ticket[8]),
       }))
-      .filter((ticket: any) => ticket.daysUntilConcert >= 0)
-    // .filter((ticket: any) => ticket.daysUntilConcert >= 0 && ticket.artist !== myUsername)
+      // .filter((ticket: any) => ticket.daysUntilConcert >= 0)
+      .filter(
+        (ticket: any) =>
+          ticket.daysUntilConcert >= 0 && ticket.artist !== myUsername,
+      )
     // console.log(myTicketsFormatted)
     setMyTickets(myTicketsFormatted)
   }
@@ -255,11 +259,12 @@ export const Marketplace: React.FC<Props> = ({
           />
           <div className={'h-[97%] pt-[20px]'}>
             {showTickets === 'Ticket Counter' ? (
-              <NftTickets
+              <TicketPurchase formattedEvents={formattedEvents} />
+            ) : showTickets === 'Your NFT Tickets' ? (
+              <YourNftTickets
                 balance={balance}
                 chainName={chainName}
                 chainId={chainId}
-                formattedEvents={formattedEvents}
                 myTickets={myTickets}
               />
             ) : showTickets === 'Your future shows' ? (
